requirements.txt
fastapi
uvicorn
pdfminer.six
pypdf
python-multipart
